Split menu.rs into module directory with 6 focused files
