--- src/ui/menu.rs
@@ -1,2762 +1,0 @@
-use std::path::{Path, PathBuf};
-use std::sync::Arc;
-use std::time::Instant;
-
-use parking_lot::Mutex;
-
-use serde::{Deserialize, Serialize};
-
-use crate::window::DisplayMode;
-
-use crate::renderer::pipelines::menu_overlay::{
-    MenuElement, ICON_CHECK, ICON_CODE, ICON_COMMENT, ICON_GEAR, ICON_GLOBE, ICON_LINK,
-    ICON_PAINTBRUSH, ICON_USER,
-};
-
-#[derive(Serialize, Deserialize)]
-struct Settings {
-    gui_scale: u32,
-    render_distance: u32,
-    simulation_distance: u32,
-}
-
-impl Default for Settings {
-    fn default() -> Self {
-        Self {
-            gui_scale: 0,
-            render_distance: 12,
-            simulation_distance: 12,
-        }
-    }
-}
-
-fn load_settings(game_dir: &Path) -> Settings {
-    let path = game_dir.join("pomc_settings.json");
-    std::fs::read_to_string(&path)
-        .ok()
-        .and_then(|s| serde_json::from_str(&s).ok())
-        .unwrap_or_default()
-}
-
-fn save_settings(game_dir: &Path, settings: &Settings) {
-    let path = game_dir.join("pomc_settings.json");
-    if let Ok(json) = serde_json::to_string(settings) {
-        let _ = std::fs::write(path, json);
-    }
-}
-
-use super::auth::{self, AuthAccount, AuthStatus};
-use super::common::{self, WHITE};
-use super::server_list::{
-    is_valid_address, ping_all_servers, PingResults, PingState, ServerEntry, ServerList,
-};
-
-#[derive(Clone, Copy, PartialEq, Eq)]
-pub enum PanoramaTheme {
-    Pomc,
-    Default,
-}
-
-struct ThemeTransition {
-    start: Instant,
-    target: PanoramaTheme,
-    reloaded: bool,
-    open_start: Option<Instant>,
-}
-
-const CLOSE_DURATION: f32 = 0.5;
-const OPEN_DURATION: f32 = 0.5;
-const STRIP_COUNT: usize = 14;
-
-pub enum MenuAction {
-    None,
-    Connect { server: String, username: String },
-    ChangeTheme(PanoramaTheme),
-    Quit,
-}
-
-pub struct MainMenuResult {
-    pub elements: Vec<MenuElement>,
-    pub action: MenuAction,
-    pub cursor_pointer: bool,
-    pub blur: f32,
-    pub clicked_button: bool,
-}
-
-pub struct MenuInput {
-    pub cursor: (f32, f32),
-    pub clicked: bool,
-    pub mouse_held: bool,
-    pub typed_chars: Vec<char>,
-    pub backspace: bool,
-    pub enter: bool,
-    pub escape: bool,
-    pub tab: bool,
-    pub f5: bool,
-    pub scroll_delta: f32,
-}
-
-const HEADER_H: f32 = 33.0;
-const ENTRY_H: f32 = 36.0;
-const ROW_W: f32 = 305.0;
-const FORM_W: f32 = 200.0;
-const BTN_GAP: f32 = 4.0;
-const TOP_BTN_W: f32 = 100.0;
-const BOT_BTN_W: f32 = 74.0;
-const SEP_H: f32 = 2.0;
-const FIELD_H: f32 = 20.0;
-
-const COL_DIM: [f32; 4] = [0.55, 0.57, 0.69, 1.0];
-const COL_DARK_DIM: [f32; 4] = [0.4, 0.42, 0.52, 1.0];
-const COL_RED: [f32; 4] = [0.88, 0.25, 0.32, 1.0];
-const COL_SEP: [f32; 4] = [1.0, 1.0, 1.0, 0.07];
-
-const FIELD_BG: [f32; 4] = [0.06, 0.07, 0.14, 0.8];
-const FIELD_BORDER: [f32; 4] = [1.0, 1.0, 1.0, 0.08];
-const FIELD_BORDER_FOCUS: [f32; 4] = [0.29, 0.87, 0.5, 0.5];
-
-const DOUBLE_CLICK_MS: u128 = 400;
-
-enum Screen {
-    Main,
-    AuthPrompt { pending: AuthPending },
-    Auth { pending: AuthPending },
-    ServerList,
-    ConfirmDelete(usize),
-    DirectConnect,
-    AddServer,
-    EditServer(usize),
-    Disconnected(String),
-    Options,
-    OptionsVideo,
-    OptionsSkinCustomization,
-    OptionsMusicSounds,
-    OptionsControls,
-    OptionsKeybinds,
-    OptionsLanguage,
-    OptionsChatSettings,
-    OptionsResourcePacks,
-    OptionsAccessibility,
-    OptionsTelemetry,
-    OptionsCredits,
-}
-
-impl Screen {
-    fn clone_screen(&self) -> Self {
-        match self {
-            Self::Main => Self::Main,
-            Self::Options => Self::Options,
-            Self::OptionsVideo => Self::OptionsVideo,
-            Self::OptionsSkinCustomization => Self::OptionsSkinCustomization,
-            Self::OptionsMusicSounds => Self::OptionsMusicSounds,
-            Self::OptionsControls => Self::OptionsControls,
-            Self::OptionsKeybinds => Self::OptionsKeybinds,
-            Self::OptionsLanguage => Self::OptionsLanguage,
-            Self::OptionsChatSettings => Self::OptionsChatSettings,
-            Self::OptionsResourcePacks => Self::OptionsResourcePacks,
-            Self::OptionsAccessibility => Self::OptionsAccessibility,
-            Self::OptionsTelemetry => Self::OptionsTelemetry,
-            Self::OptionsCredits => Self::OptionsCredits,
-            Self::ServerList => Self::ServerList,
-            Self::DirectConnect => Self::DirectConnect,
-            Self::AddServer => Self::AddServer,
-            Self::AuthPrompt { pending } => Self::AuthPrompt { pending: *pending },
-            Self::Auth { pending } => Self::Auth { pending: *pending },
-            Self::ConfirmDelete(i) => Self::ConfirmDelete(*i),
-            Self::EditServer(i) => Self::EditServer(*i),
-            Self::Disconnected(s) => Self::Disconnected(s.clone()),
-        }
-    }
-}
-
-#[derive(Clone, Copy)]
-enum AuthPending {
-    None,
-    Singleplayer,
-    Multiplayer,
-}
-
-pub struct MainMenu {
-    username: String,
-    screen: Screen,
-    server_list: ServerList,
-    selected_server: Option<usize>,
-    edit_name: String,
-    edit_address: String,
-    last_mp_ip: String,
-    ping_results: PingResults,
-    rt: Arc<tokio::runtime::Runtime>,
-    links_open: bool,
-    theme_open: bool,
-    theme: PanoramaTheme,
-    transition: Option<ThemeTransition>,
-    scroll_offset: f32,
-    focused_field: Option<u8>,
-    cursor_blink: Instant,
-    last_click_time: Instant,
-    last_click_index: Option<usize>,
-    auth_status: Arc<Mutex<AuthStatus>>,
-    auth_account: Option<AuthAccount>,
-    cache_file: PathBuf,
-    pub gui_scale_setting: u32,
-    pub render_distance: u32,
-    pub simulation_distance: u32,
-    pub display_mode: DisplayMode,
-    active_slider: Option<&'static str>,
-    settings_dir: PathBuf,
-    menu_open_time: Option<Instant>,
-}
-
-impl MainMenu {
-    pub fn new(game_dir: &Path, rt: Arc<tokio::runtime::Runtime>) -> Self {
-        let server_list = ServerList::load(game_dir);
-        let ping_results: PingResults = Default::default();
-        ping_all_servers(&rt, &server_list.servers, &ping_results);
-        let cache_file = game_dir.join("auth_cache.json");
-        let auth_account = auth::try_restore_cached(&cache_file);
-        let username = auth_account
-            .as_ref()
-            .map(|a| a.username.clone())
-            .unwrap_or_else(|| "Steve".into());
-        let settings = load_settings(game_dir);
-        Self {
-            username,
-            screen: Screen::Main,
-            server_list,
-            selected_server: None,
-            edit_name: String::new(),
-            edit_address: String::new(),
-            last_mp_ip: String::new(),
-            ping_results,
-            rt,
-            links_open: false,
-            theme_open: false,
-            theme: PanoramaTheme::Pomc,
-            transition: None,
-            scroll_offset: 0.0,
-            focused_field: None,
-            cursor_blink: Instant::now(),
-            last_click_time: Instant::now(),
-            last_click_index: None,
-            auth_status: Arc::new(Mutex::new(AuthStatus::Idle)),
-            auth_account,
-            cache_file,
-            gui_scale_setting: settings.gui_scale,
-            render_distance: settings.render_distance,
-            simulation_distance: settings.simulation_distance,
-            display_mode: DisplayMode::Windowed,
-            active_slider: None,
-            settings_dir: game_dir.to_path_buf(),
-            menu_open_time: None,
-        }
-    }
-
-    fn save_settings(&self) {
-        save_settings(
-            &self.settings_dir,
-            &Settings {
-                gui_scale: self.gui_scale_setting,
-                render_distance: self.render_distance,
-                simulation_distance: self.simulation_distance,
-            },
-        );
-    }
-
-    pub fn open_options(&mut self) {
-        self.screen = Screen::Options;
-    }
-
-    pub fn is_options_screen(&self) -> bool {
-        matches!(
-            self.screen,
-            Screen::Options
-                | Screen::OptionsVideo
-                | Screen::OptionsSkinCustomization
-                | Screen::OptionsMusicSounds
-                | Screen::OptionsControls
-                | Screen::OptionsKeybinds
-                | Screen::OptionsLanguage
-                | Screen::OptionsChatSettings
-                | Screen::OptionsResourcePacks
-                | Screen::OptionsAccessibility
-                | Screen::OptionsTelemetry
-                | Screen::OptionsCredits
-        )
-    }
-
-    pub fn start_transition_open(&mut self) {
-        if let Some(ref mut tr) = self.transition {
-            tr.open_start = Some(Instant::now());
-        }
-    }
-
-    pub fn show_disconnect(&mut self, reason: String) {
-        self.screen = Screen::Disconnected(reason);
-    }
-
-    pub fn build(
-        &mut self,
-        screen_w: f32,
-        screen_h: f32,
-        input: &MenuInput,
-        text_width_fn: impl Fn(&str, f32) -> f32,
-    ) -> MainMenuResult {
-        match self.screen {
-            Screen::Main => self.build_main(screen_w, screen_h, input, text_width_fn),
-            Screen::AuthPrompt { .. } => {
-                self.build_auth_prompt(screen_w, screen_h, input, &text_width_fn)
-            }
-            Screen::Auth { .. } => self.build_auth(screen_w, screen_h, input, &text_width_fn),
-            Screen::ServerList => self.build_server_list(screen_w, screen_h, input, &text_width_fn),
-            Screen::ConfirmDelete(_) => {
-                self.build_confirm_delete(screen_w, screen_h, input, &text_width_fn)
-            }
-            Screen::DirectConnect => {
-                self.build_direct_connect(screen_w, screen_h, input, &text_width_fn)
-            }
-            Screen::AddServer | Screen::EditServer(_) => {
-                self.build_edit_server(screen_w, screen_h, input, &text_width_fn)
-            }
-            Screen::Disconnected(_) => {
-                self.build_disconnected(screen_w, screen_h, input, &text_width_fn)
-            }
-            Screen::Options => self.build_options(screen_w, screen_h, input),
-            Screen::OptionsVideo => self.build_options_video(screen_w, screen_h, input),
-            Screen::OptionsSkinCustomization => self.build_options_stub(
-                screen_w,
-                screen_h,
-                input,
-                "Skin Customization",
-                Screen::Options,
-            ),
-            Screen::OptionsMusicSounds => self.build_options_stub(
-                screen_w,
-                screen_h,
-                input,
-                "Music & Sounds",
-                Screen::Options,
-            ),
-            Screen::OptionsControls => self.build_options_controls(screen_w, screen_h, input),
-            Screen::OptionsKeybinds => self.build_options_stub(
-                screen_w,
-                screen_h,
-                input,
-                "Keybinds",
-                Screen::OptionsControls,
-            ),
-            Screen::OptionsLanguage => {
-                self.build_options_stub(screen_w, screen_h, input, "Language", Screen::Options)
-            }
-            Screen::OptionsChatSettings => {
-                self.build_options_stub(screen_w, screen_h, input, "Chat Settings", Screen::Options)
-            }
-            Screen::OptionsResourcePacks => self.build_options_stub(
-                screen_w,
-                screen_h,
-                input,
-                "Resource Packs",
-                Screen::Options,
-            ),
-            Screen::OptionsAccessibility => self.build_options_stub(
-                screen_w,
-                screen_h,
-                input,
-                "Accessibility Settings",
-                Screen::Options,
-            ),
-            Screen::OptionsTelemetry => self.build_options_stub(
-                screen_w,
-                screen_h,
-                input,
-                "Telemetry Data",
-                Screen::Options,
-            ),
-            Screen::OptionsCredits => self.build_options_stub(
-                screen_w,
-                screen_h,
-                input,
-                "Credits & Attribution",
-                Screen::Options,
-            ),
-        }
-    }
-
-    #[allow(clippy::too_many_lines)]
-    fn build_main(
-        &mut self,
-        screen_w: f32,
-        screen_h: f32,
-        input: &MenuInput,
-        text_width_fn: impl Fn(&str, f32) -> f32,
-    ) -> MainMenuResult {
-        let gs = crate::ui::hud::gui_scale(screen_w, screen_h, self.gui_scale_setting);
-        let cursor = input.cursor;
-        let clicked = input.clicked;
-
-        let mut elements = Vec::new();
-        let mut action = MenuAction::None;
-        let mut any_hovered = false;
-        let mut any_clicked = false;
-
-        let anim_t = self
-            .menu_open_time
-            .get_or_insert_with(Instant::now)
-            .elapsed()
-            .as_secs_f32();
-        let panel_t = ease_out_cubic((anim_t / 2.0).min(1.0));
-
-        let accent: [f32; 4] = [0.29, 0.87, 0.5, 1.0];
-        let glass: [f32; 4] = [0.07, 0.08, 0.16, 0.55];
-        let glass_hover: [f32; 4] = [0.12, 0.14, 0.25, 0.65];
-        let text_col: [f32; 4] = [0.89, 0.90, 0.96, 0.85];
-        let text_bright: [f32; 4] = [0.94, 0.95, 0.98, 1.0];
-        let text_dim: [f32; 4] = [0.53, 0.56, 0.69, 0.6];
-        let border: [f32; 4] = [1.0, 1.0, 1.0, 0.05];
-
-        struct BtnDef {
-            label: &'static str,
-            id: u8,
-        }
-        let buttons = [
-            BtnDef {
-                label: "Singleplayer",
-                id: 0,
-            },
-            BtnDef {
-                label: "Multiplayer",
-                id: 1,
-            },
-            BtnDef {
-                label: "Quit Game",
-                id: 2,
-            },
-        ];
-
-        let s = (screen_h / 400.0).max(1.0);
-        let panel_w = (260.0 * s).min(screen_w * 0.4);
-        let panel_pad = 28.0 * s;
-        let panel_r = 14.0 * s;
-        let accent_bar_h = 3.0 * s;
-        let title_size = 40.0 * s;
-        let sub_size = 9.0 * s;
-        let content_w = panel_w - panel_pad * 2.0;
-        let btn_h = 36.0 * s;
-        let btn_gap = 5.0 * s;
-        let btn_r = 8.0 * s;
-        let font_size = 11.0 * s;
-        let accent_w = 3.0 * s;
-        let icon_size = 28.0 * s;
-        let icon_gap = 6.0 * s;
-
-        let header_h = accent_bar_h + 14.0 * s + title_size + 4.0 * s + sub_size + 18.0 * s;
-        let btns_total = buttons.len() as f32 * (btn_h + btn_gap) - btn_gap;
-        let panel_h =
-            (panel_pad + header_h + 1.0 + 16.0 * s + btns_total + 16.0 * s + icon_size + panel_pad)
-                .min(screen_h * 0.9);
-        let panel_margin = (screen_w * 0.06).max(12.0);
-        let panel_start_x = -panel_w;
-        let panel_final_x = panel_margin;
-        let panel_x = panel_start_x + (panel_final_x - panel_start_x) * panel_t;
-        let panel_y = (screen_h - panel_h) / 2.0;
-        let btn_x = panel_x + panel_pad;
-
-        elements.push(MenuElement::FrostedRect {
-            x: panel_x,
-            y: panel_y,
-            w: panel_w,
-            h: panel_h,
-            corner_radius: panel_r,
-            tint: [0.055, 0.06, 0.13, 0.72],
-        });
-
-        let mut cy = panel_y + panel_pad;
-
-        elements.push(MenuElement::Rect {
-            x: btn_x,
-            y: cy,
-            w: 50.0 * s,
-            h: accent_bar_h,
-            corner_radius: accent_bar_h * 0.5,
-            color: [accent[0], accent[1], accent[2], 0.7],
-        });
-        cy += accent_bar_h + 14.0 * s;
-
-        let pomc_w = text_width_fn("POMC", title_size);
-        elements.push(MenuElement::Text {
-            x: btn_x,
-            y: cy,
-            text: "POMC".into(),
-            scale: title_size,
-            color: [0.94, 0.96, 0.99, 0.95],
-            centered: false,
-        });
-
-        let sub_x = btn_x + pomc_w + 8.0 * s;
-        let sub_y1 = cy + title_size - sub_size * 2.0 - 4.0 * s;
-        let sub_y2 = cy + title_size - sub_size - 1.0 * s;
-        let badge_t = ((anim_t - 2.0) / 0.3).clamp(0.0, 1.0);
-        let badge_scale = ease_out_cubic(badge_t);
-
-        let rust_w = text_width_fn("Rust", sub_size);
-        let badge_pad_x = 5.0 * s;
-        let badge_pad_y = 2.5 * s;
-        let badge_w = rust_w + badge_pad_x * 2.0;
-        let badge_h = sub_size + badge_pad_y * 2.0;
-        let badge_r = badge_h * 0.5;
-
-        if badge_t < 1.0 {
-            elements.push(MenuElement::Text {
-                x: sub_x,
-                y: sub_y1,
-                text: "Java".into(),
-                scale: sub_size,
-                color: text_dim,
-                centered: false,
-            });
-        }
-
-        if badge_t > 0.0 {
-            let bx = sub_x - badge_pad_x;
-            let by = sub_y1 - badge_pad_y;
-            let bw = badge_w * badge_scale;
-            elements.push(MenuElement::Rect {
-                x: bx,
-                y: by,
-                w: bw,
-                h: badge_h,
-                corner_radius: badge_r,
-                color: [accent[0], accent[1], accent[2], 0.9 * badge_scale],
-            });
-            if badge_t >= 0.5 {
-                let text_a = ((badge_t - 0.5) / 0.5).min(1.0);
-                elements.push(MenuElement::Text {
-                    x: sub_x,
-                    y: sub_y1,
-                    text: "Rust".into(),
-                    scale: sub_size,
-                    color: [0.05, 0.05, 0.1, text_a],
-                    centered: false,
-                });
-            }
-        }
-
-        elements.push(MenuElement::Text {
-            x: sub_x,
-            y: sub_y2,
-            text: "Edition".into(),
-            scale: sub_size,
-            color: text_dim,
-            centered: false,
-        });
-        cy += title_size + 18.0 * s;
-
-        elements.push(MenuElement::Rect {
-            x: btn_x,
-            y: cy,
-            w: content_w,
-            h: 1.0,
-            corner_radius: 0.5,
-            color: border,
-        });
-        cy += 1.0 + 16.0 * s;
-
-        for (i, def) in buttons.iter().enumerate() {
-            let by = cy + i as f32 * (btn_h + btn_gap);
-            let rect = [btn_x, by, content_w, btn_h];
-            let hovered = common::hit_test(cursor, rect);
-            any_hovered |= hovered;
-
-            elements.push(MenuElement::Rect {
-                x: rect[0],
-                y: rect[1],
-                w: rect[2],
-                h: rect[3],
-                corner_radius: btn_r,
-                color: if hovered { glass_hover } else { glass },
-            });
-
-            let bar_margin = btn_h * 0.18;
-            elements.push(MenuElement::Rect {
-                x: rect[0],
-                y: rect[1] + bar_margin,
-                w: accent_w,
-                h: rect[3] - bar_margin * 2.0,
-                corner_radius: accent_w * 0.5,
-                color: [
-                    accent[0],
-                    accent[1],
-                    accent[2],
-                    if hovered { 0.9 } else { 0.12 },
-                ],
-            });
-
-            elements.push(MenuElement::Text {
-                x: rect[0] + 18.0 * s,
-                y: rect[1] + (rect[3] - font_size) / 2.0,
-                text: def.label.into(),
-                scale: font_size,
-                color: if hovered { text_bright } else { text_col },
-                centered: false,
-            });
-
-            if clicked && hovered {
-                any_clicked = true;
-                if def.id == 2 {
-                    action = MenuAction::Quit;
-                } else if self.auth_account.is_some() {
-                    match def.id {
-                        0 => {}
-                        1 => {
-                            self.screen = Screen::ServerList;
-                            self.scroll_offset = 0.0;
-                            self.selected_server = None;
-                        }
-                        _ => {}
-                    }
-                } else {
-                    let pending = match def.id {
-                        0 => AuthPending::Singleplayer,
-                        _ => AuthPending::Multiplayer,
-                    };
-                    self.screen = Screen::AuthPrompt { pending };
-                }
-            }
-        }
-
-        let icon_area_y = panel_y + panel_h - panel_pad - icon_size;
-        let icon_r = 7.0 * s;
-        let icon_scale = 13.0 * s;
-        let drop_style = DropdownStyle::new(gs);
-
-        let bottom_icons: [(f32, char); 4] = [
-            (btn_x, ICON_USER),
-            (btn_x + icon_size + icon_gap, ICON_LINK),
-            (btn_x + content_w - icon_size, ICON_GEAR),
-            (
-                btn_x + content_w - icon_size * 2.0 - icon_gap,
-                ICON_PAINTBRUSH,
-            ),
-        ];
-
-        for &(bx, icon) in &bottom_icons {
-            let rect = [bx, icon_area_y, icon_size, icon_size];
-            let hovered = common::hit_test(cursor, rect);
-            any_hovered |= hovered;
-
-            elements.push(MenuElement::Rect {
-                x: bx,
-                y: icon_area_y,
-                w: icon_size,
-                h: icon_size,
-                corner_radius: icon_r,
-                color: if hovered {
-                    glass_hover
-                } else {
-                    [0.0, 0.0, 0.0, 0.0]
-                },
-            });
-            elements.push(MenuElement::Icon {
-                x: bx + icon_size / 2.0,
-                y: icon_area_y + icon_size / 2.0,
-                icon,
-                scale: icon_scale,
-                color: if hovered { text_bright } else { text_dim },
-            });
-
-            if clicked && hovered {
-                any_clicked = true;
-                match icon {
-                    ICON_USER if self.auth_account.is_none() => {
-                        self.screen = Screen::AuthPrompt {
-                            pending: AuthPending::None,
-                        };
-                    }
-                    ICON_LINK => {
-                        self.links_open = !self.links_open;
-                        if self.links_open {
-                            self.theme_open = false;
-                        }
-                    }
-                    ICON_GEAR => {
-                        self.open_options();
-                    }
-                    ICON_PAINTBRUSH => {
-                        self.theme_open = !self.theme_open;
-                        if self.theme_open {
-                            self.links_open = false;
-                        }
-                    }
-                    _ => {}
-                }
-            }
-        }
-
-        if self.links_open {
-            let anchor_x = btn_x + icon_size + icon_gap;
-            let drop_w = 140.0 * s;
-            let drop_x = anchor_x;
-            let drop_y = icon_area_y - 2.0 * s;
-            let links: [(&str, char, &str); 3] = [
-                ("Website", ICON_GLOBE, "https://website.com"),
-                ("Discord", ICON_COMMENT, "https://discord.gg/ucBA55bHPR"),
-                ("GitHub", ICON_CODE, "https://github.com"),
-            ];
-            let total_h = links.len() as f32 * drop_style.item_h;
-            let drop_y_top = drop_y - total_h;
-            drop_style.draw_background(&mut elements, drop_x, drop_y_top, drop_w, total_h);
-            let mut clicked_inside = false;
-            for (i, (label, icon, url)) in links.iter().enumerate() {
-                let item = drop_style.draw_item(
-                    &mut elements,
-                    &mut any_hovered,
-                    cursor,
-                    drop_x,
-                    drop_y_top,
-                    drop_w,
-                    i,
-                    links.len(),
-                    label,
-                    Some((*icon, [0.6, 0.7, 0.85, 0.8])),
-                    text_bright,
-                    text_col,
-                );
-                if item {
-                    clicked_inside = true;
-                }
-                if clicked && item {
-                    let _ = open::that(url);
-                    self.links_open = false;
-                }
-            }
-            if dismiss_dropdown(
-                cursor,
-                clicked,
-                clicked_inside,
-                [drop_x, drop_y_top, drop_w, total_h],
-                [anchor_x, icon_area_y, icon_size, icon_size],
-            ) {
-                self.links_open = false;
-            }
-        }
-
-        if self.theme_open {
-            let anchor_x = btn_x + content_w - icon_size * 2.0 - icon_gap;
-            let drop_w = 120.0 * s;
-            let drop_x = anchor_x + icon_size - drop_w;
-            let drop_y = icon_area_y - 2.0 * s;
-            let themes: [(&str, PanoramaTheme); 2] = [
-                ("POMC", PanoramaTheme::Pomc),
-                ("Default", PanoramaTheme::Default),
-            ];
-            let total_h = themes.len() as f32 * drop_style.item_h;
-            let drop_y_top = drop_y - total_h;
-            drop_style.draw_background(&mut elements, drop_x, drop_y_top, drop_w, total_h);
-            let mut clicked_inside = false;
-            for (i, (label, theme_val)) in themes.iter().enumerate() {
-                let selected = self.theme == *theme_val;
-                let check = if selected {
-                    Some((ICON_CHECK, [0.39, 0.71, 1.0, 0.9]))
-                } else {
-                    None
-                };
-                let text_c = if selected {
-                    [0.39, 0.71, 1.0, 0.9]
-                } else {
-                    text_col
-                };
-                let item = drop_style.draw_item(
-                    &mut elements,
-                    &mut any_hovered,
-                    cursor,
-                    drop_x,
-                    drop_y_top,
-                    drop_w,
-                    i,
-                    themes.len(),
-                    label,
-                    check,
-                    text_bright,
-                    text_c,
-                );
-                if item {
-                    clicked_inside = true;
-                }
-                if clicked && item && !selected {
-                    self.transition = Some(ThemeTransition {
-                        start: Instant::now(),
-                        target: *theme_val,
-                        reloaded: false,
-                        open_start: None,
-                    });
-                    self.theme_open = false;
-                } else if clicked && item {
-                    self.theme_open = false;
-                }
-            }
-            if dismiss_dropdown(
-                cursor,
-                clicked,
-                clicked_inside,
-                [drop_x, drop_y_top, drop_w, total_h],
-                [anchor_x, icon_area_y, icon_size, icon_size],
-            ) {
-                self.theme_open = false;
-            }
-        }
-
-        let footer_size = 8.0 * s;
-        let footer_pad = 8.0 * s;
-        let footer_y = screen_h - footer_pad - footer_size;
-        let footer_col = [0.4, 0.45, 0.6, 0.2];
-        elements.push(MenuElement::Text {
-            x: footer_pad,
-            y: footer_y,
-            text: "1.21.11".into(),
-            scale: footer_size,
-            color: footer_col,
-            centered: false,
-        });
-        let copy = "POMC early dev";
-        let copy_w = text_width_fn(copy, footer_size);
-        elements.push(MenuElement::Text {
-            x: screen_w - footer_pad - copy_w,
-            y: footer_y,
-            text: copy.into(),
-            scale: footer_size,
-            color: footer_col,
-            centered: false,
-        });
-
-        if let Some(ref mut tr) = self.transition {
-            let close_t = (tr.start.elapsed().as_secs_f32() / CLOSE_DURATION).min(1.0);
-            if close_t >= 1.0 && !tr.reloaded {
-                tr.reloaded = true;
-                self.theme = tr.target;
-                action = MenuAction::ChangeTheme(tr.target);
-            }
-            let open_t = tr
-                .open_start
-                .map(|s| (s.elapsed().as_secs_f32() / OPEN_DURATION).min(1.0))
-                .unwrap_or(0.0);
-            emit_transition_strips(&mut elements, screen_w, screen_h, close_t, open_t);
-            if open_t >= 1.0 {
-                self.transition = None;
-            }
-        }
-
-        MainMenuResult {
-            elements,
-            action,
-            cursor_pointer: any_hovered,
-            blur: 1.0,
-            clicked_button: any_clicked,
-        }
-    }
-
-    pub fn set_launch_auth(&mut self, username: String, uuid: uuid::Uuid, access_token: String) {
-        self.username = username.clone();
-        self.auth_account = Some(AuthAccount {
-            username,
-            uuid,
-            access_token,
-        });
-    }
-
-    pub fn auth_account(&self) -> Option<&AuthAccount> {
-        self.auth_account.as_ref()
-    }
-
-    fn build_auth_prompt(
-        &mut self,
-        screen_w: f32,
-        screen_h: f32,
-        input: &MenuInput,
-        _text_width_fn: &dyn Fn(&str, f32) -> f32,
-    ) -> MainMenuResult {
-        let Screen::AuthPrompt { pending } = self.screen else {
-            return empty_result(2.0);
-        };
-
-        if input.escape {
-            self.screen = Screen::Main;
-            return empty_result(2.0);
-        }
-
-        let gs = crate::ui::hud::gui_scale(screen_w, screen_h, self.gui_scale_setting);
-        let title_size = 18.0 * gs;
-        let body_size = 10.0 * gs;
-        let btn_w = 180.0 * gs;
-        let btn_h = 30.0 * gs;
-        let gap = 10.0 * gs;
-        let cx = screen_w / 2.0;
-        let dim: [f32; 4] = [0.7, 0.7, 0.7, 0.8];
-
-        let lines = [
-            "You need to sign in with your Microsoft account.",
-            "",
-            "A browser tab will open where you can sign in.",
-            "Once complete, the client will detect it automatically.",
-            "",
-            "In the future, a launcher will handle authentication.",
-            "For now, we use a temporary sign-in method.",
-        ];
-
-        let text_h = lines.len() as f32 * (body_size + 3.0 * gs);
-        let total_h = title_size + gap * 2.0 + text_h + gap * 2.0 + btn_h + gap + btn_h;
-        let mut y = (screen_h - total_h) / 2.0;
-
-        let mut elements = Vec::new();
-        let mut any_hovered = false;
-
-        elements.push(MenuElement::Text {
-            x: cx,
-            y,
-            text: "Sign In Required".into(),
-            scale: title_size,
-            color: WHITE,
-            centered: true,
-        });
-        y += title_size + gap * 2.0;
-
-        for line in &lines {
-            if !line.is_empty() {
-                elements.push(MenuElement::Text {
-                    x: cx,
-                    y,
-                    text: (*line).into(),
-                    scale: body_size,
-                    color: dim,
-                    centered: true,
-                });
-            }
-            y += body_size + 3.0 * gs;
-        }
-        y += gap;
-
-        if push_button(
-            &mut elements,
-            &mut any_hovered,
-            input.cursor,
-            cx - btn_w / 2.0,
-            y,
-            btn_w,
-            btn_h,
-            gs,
-            "Sign in with Microsoft",
-            true,
-        ) && input.clicked
-        {
-            self.screen = Screen::Auth { pending };
-            auth::spawn_auth(
-                &self.rt,
-                Arc::clone(&self.auth_status),
-                self.cache_file.clone(),
-            );
-        }
-        y += btn_h + gap;
-
-        if push_button(
-            &mut elements,
-            &mut any_hovered,
-            input.cursor,
-            cx - btn_w / 2.0,
-            y,
-            btn_w,
-            btn_h,
-            gs,
-            "Back",
-            true,
-        ) && input.clicked
-        {
-            self.screen = Screen::Main;
-        }
-
-        MainMenuResult {
-            elements,
-            action: MenuAction::None,
-            cursor_pointer: any_hovered,
-            blur: 2.0,
-            clicked_button: false,
-        }
-    }
-
-    fn cancel_auth(&mut self) {
-        self.screen = Screen::Main;
-        *self.auth_status.lock() = AuthStatus::Idle;
-    }
-
-    fn build_auth(
-        &mut self,
-        screen_w: f32,
-        screen_h: f32,
-        input: &MenuInput,
-        _text_width_fn: &dyn Fn(&str, f32) -> f32,
-    ) -> MainMenuResult {
-        let Screen::Auth { pending } = self.screen else {
-            return empty_result(2.0);
-        };
-
-        let gs = crate::ui::hud::gui_scale(screen_w, screen_h, self.gui_scale_setting);
-        let title_size = 18.0 * gs;
-        let body_size = 11.0 * gs;
-        let btn_w = 160.0 * gs;
-        let btn_h = 30.0 * gs;
-        let gap = 12.0 * gs;
-        let cx = screen_w / 2.0;
-        let status_color: [f32; 4] = [0.8, 0.8, 0.8, 0.9];
-
-        let mut elements = Vec::new();
-        let mut any_hovered = false;
-
-        let status = self.auth_status.lock();
-        match &*status {
-            AuthStatus::Idle | AuthStatus::OpeningBrowser => {
-                elements.push(MenuElement::Text {
-                    x: cx,
-                    y: (screen_h - body_size) / 2.0,
-                    text: "Opening browser...".into(),
-                    scale: body_size,
-                    color: status_color,
-                    centered: true,
-                });
-            }
-            AuthStatus::WaitingForBrowser => {
-                drop(status);
-
-                let total_h = title_size + gap + body_size + gap * 2.0 + btn_h;
-                let mut y = (screen_h - total_h) / 2.0;
-
-                elements.push(MenuElement::Text {
-                    x: cx,
-                    y,
-                    text: "Sign in with Microsoft".into(),
-                    scale: title_size,
-                    color: WHITE,
-                    centered: true,
-                });
-                y += title_size + gap;
-
-                elements.push(MenuElement::Text {
-                    x: cx,
-                    y,
-                    text: "Complete sign-in in your browser...".into(),
-                    scale: body_size,
-                    color: status_color,
-                    centered: true,
-                });
-                y += body_size + gap * 2.0;
-
-                if push_button(
-                    &mut elements,
-                    &mut any_hovered,
-                    input.cursor,
-                    cx - btn_w / 2.0,
-                    y,
-                    btn_w,
-                    btn_h,
-                    gs,
-                    "Cancel",
-                    true,
-                ) && input.clicked
-                {
-                    self.cancel_auth();
-                    return empty_result(2.0);
-                }
-
-                return MainMenuResult {
-                    elements,
-                    action: MenuAction::None,
-                    cursor_pointer: any_hovered,
-                    blur: 2.0,
-                    clicked_button: false,
-                };
-            }
-            AuthStatus::Exchanging => {
-                elements.push(MenuElement::Text {
-                    x: cx,
-                    y: (screen_h - body_size) / 2.0,
-                    text: "Logging in...".into(),
-                    scale: body_size,
-                    color: status_color,
-                    centered: true,
-                });
-            }
-            AuthStatus::Success(_) => {
-                drop(status);
-                let old = std::mem::replace(&mut *self.auth_status.lock(), AuthStatus::Idle);
-                if let AuthStatus::Success(account) = old {
-                    self.username = account.username.clone();
-                    self.auth_account = Some(account);
-                }
-
-                match pending {
-                    AuthPending::None | AuthPending::Singleplayer => self.screen = Screen::Main,
-                    AuthPending::Multiplayer => {
-                        self.screen = Screen::ServerList;
-                        self.scroll_offset = 0.0;
-                        self.selected_server = None;
-                    }
-                }
-                return empty_result(2.0);
-            }
-            AuthStatus::Failed(err) => {
-                let err = err.clone();
-                drop(status);
-
-                let total_h = title_size + gap + body_size + gap * 2.0 + btn_h;
-                let mut y = (screen_h - total_h) / 2.0;
-
-                elements.push(MenuElement::Text {
-                    x: cx,
-                    y,
-                    text: "Authentication Failed".into(),
-                    scale: title_size,
-                    color: [1.0, 0.4, 0.4, 1.0],
-                    centered: true,
-                });
-                y += title_size + gap;
-
-                elements.push(MenuElement::Text {
-                    x: cx,
-                    y,
-                    text: err,
-                    scale: body_size,
-                    color: [0.85, 0.85, 0.85, 0.9],
-                    centered: true,
-                });
-                y += body_size + gap * 2.0;
-
-                if push_button(
-                    &mut elements,
-                    &mut any_hovered,
-                    input.cursor,
-                    cx - btn_w / 2.0,
-                    y,
-                    btn_w,
-                    btn_h,
-                    gs,
-                    "Back",
-                    true,
-                ) && input.clicked
-                {
-                    self.cancel_auth();
-                    return empty_result(2.0);
-                }
-
-                return MainMenuResult {
-                    elements,
-                    action: MenuAction::None,
-                    cursor_pointer: any_hovered,
-                    blur: 2.0,
-                    clicked_button: false,
-                };
-            }
-        }
-        drop(status);
-
-        let btn_y = screen_h / 2.0 + gap * 2.0;
-        if push_button(
-            &mut elements,
-            &mut any_hovered,
-            input.cursor,
-            cx - btn_w / 2.0,
-            btn_y,
-            btn_w,
-            btn_h,
-            gs,
-            "Cancel",
-            true,
-        ) && input.clicked
-        {
-            self.cancel_auth();
-        }
-
-        MainMenuResult {
-            elements,
-            action: MenuAction::None,
-            cursor_pointer: any_hovered,
-            blur: 2.0,
-            clicked_button: false,
-        }
-    }
-
-    fn build_server_list(
-        &mut self,
-        screen_w: f32,
-        screen_h: f32,
-        input: &MenuInput,
-        text_width_fn: &dyn Fn(&str, f32) -> f32,
-    ) -> MainMenuResult {
-        let gs = crate::ui::hud::gui_scale(screen_w, screen_h, self.gui_scale_setting);
-        let header_h = HEADER_H * gs;
-        let sep_h = SEP_H * gs;
-        let entry_h = ENTRY_H * gs;
-        let row_w = ROW_W * gs;
-        let gap = BTN_GAP * gs;
-        let fs = common::FONT_SIZE * gs;
-        let btn_h = common::BTN_H * gs;
-        let top_w = TOP_BTN_W * gs;
-        let bot_w = BOT_BTN_W * gs;
-        let cursor = input.cursor;
-        let clicked = input.clicked;
-
-        let footer_h = btn_h * 2.0 + gap * 3.0;
-        let list_top = header_h + sep_h + gap;
-        let list_bottom = screen_h - footer_h - sep_h;
-        let list_h = list_bottom - list_top;
-
-        let mut elements = Vec::new();
-        let mut action = MenuAction::None;
-        let mut any_hovered = false;
-
-        if input.f5 {
-            self.refresh_servers();
-        }
-        if input.escape {
-            self.screen = Screen::Main;
-            return MainMenuResult {
-                elements: Vec::new(),
-                action: MenuAction::None,
-                cursor_pointer: false,
-                blur: 1.0,
-                clicked_button: false,
-            };
-        }
-
-        elements.push(MenuElement::Text {
-            x: screen_w / 2.0,
-            y: (header_h - fs) / 2.0,
-            text: "Multiplayer".into(),
-            scale: fs,
-            color: WHITE,
-            centered: true,
-        });
-
-        push_separator(&mut elements, 0.0, header_h, screen_w, sep_h);
-        push_separator(&mut elements, 0.0, list_bottom, screen_w, sep_h);
-
-        let total_content = self.server_list.servers.len() as f32 * entry_h;
-        let max_scroll = (total_content - list_h).max(0.0);
-        self.scroll_offset =
-            (self.scroll_offset - input.scroll_delta * entry_h).clamp(0.0, max_scroll);
-
-        let list_cx = screen_w / 2.0;
-        let list_left = list_cx - row_w / 2.0;
-        let ping_results = self.ping_results.read().clone();
-
-        for (i, server) in self.server_list.servers.iter().enumerate() {
-            let ey = list_top + i as f32 * entry_h - self.scroll_offset;
-            if ey + entry_h < list_top || ey > list_bottom {
-                continue;
-            }
-
-            let selected = self.selected_server == Some(i);
-            let rect = [list_left, ey, row_w, entry_h];
-            let hovered =
-                common::hit_test(cursor, rect) && cursor.1 >= list_top && cursor.1 <= list_bottom;
-            any_hovered |= hovered;
-
-            if selected || hovered {
-                elements.push(MenuElement::Rect {
-                    x: rect[0],
-                    y: rect[1],
-                    w: rect[2],
-                    h: rect[3],
-                    corner_radius: 0.0,
-                    color: if selected {
-                        [1.0, 1.0, 1.0, 0.12]
-                    } else {
-                        [1.0, 1.0, 1.0, 0.06]
-                    },
-                });
-            }
-            if selected {
-                push_outline(&mut elements, rect[0], rect[1], rect[2], rect[3], gs);
-            }
-
-            let text_x = rect[0] + 3.0 * gs;
-            let name_y = rect[1] + 1.0 * gs;
-            elements.push(MenuElement::Text {
-                x: text_x,
-                y: name_y,
-                text: server.name.clone(),
-                scale: fs,
-                color: WHITE,
-                centered: false,
-            });
-
-            let motd_y = name_y + fs + 3.0 * gs;
-            push_server_status(
-                &mut elements,
-                &ping_results,
-                &server.address,
-                text_x,
-                motd_y,
-                &rect,
-                fs,
-                gs,
-                text_width_fn,
-            );
-
-            if clicked && hovered {
-                let now = Instant::now();
-                let is_double = self.last_click_index == Some(i)
-                    && now.duration_since(self.last_click_time).as_millis() < DOUBLE_CLICK_MS;
-
-                if is_double {
-                    action = MenuAction::Connect {
-                        server: server.address.clone(),
-                        username: self.username.clone(),
-                    };
-                } else {
-                    self.selected_server = Some(i);
-                    self.last_click_time = now;
-                    self.last_click_index = Some(i);
-                }
-            }
-        }
-
-        if self.server_list.servers.is_empty() {
-            elements.push(MenuElement::Text {
-                x: screen_w / 2.0,
-                y: list_top + 40.0 * gs,
-                text: "No servers added".into(),
-                scale: fs,
-                color: COL_DIM,
-                centered: true,
-            });
-        }
-
-        let has_sel = self.selected_server.is_some();
-        let footer_y = list_bottom + sep_h + gap;
-
-        let row1_w = top_w * 3.0 + gap * 2.0;
-        let row1_x = (screen_w - row1_w) / 2.0;
-
-        if push_button(
-            &mut elements,
-            &mut any_hovered,
-            cursor,
-            row1_x,
-            footer_y,
-            top_w,
-            btn_h,
-            gs,
-            "Join Server",
-            has_sel,
-        ) && clicked
-        {
-            if let Some(idx) = self.selected_server {
-                if let Some(server) = self.server_list.servers.get(idx) {
-                    action = MenuAction::Connect {
-                        server: server.address.clone(),
-                        username: self.username.clone(),
-                    };
-                }
-            }
-        }
-        if push_button(
-            &mut elements,
-            &mut any_hovered,
-            cursor,
-            row1_x + top_w + gap,
-            footer_y,
-            top_w,
-            btn_h,
-            gs,
-            "Direct Connect",
-            true,
-        ) && clicked
-        {
-            self.edit_address = self.last_mp_ip.clone();
-            self.focused_field = Some(0);
-            self.cursor_blink = Instant::now();
-            self.screen = Screen::DirectConnect;
-        }
-        if push_button(
-            &mut elements,
-            &mut any_hovered,
-            cursor,
-            row1_x + (top_w + gap) * 2.0,
-            footer_y,
-            top_w,
-            btn_h,
-            gs,
-            "Add Server",
-            true,
-        ) && clicked
-        {
-            self.edit_name.clear();
-            self.edit_address.clear();
-            self.focused_field = Some(0);
-            self.cursor_blink = Instant::now();
-            self.screen = Screen::AddServer;
-        }
-
-        let row2_y = footer_y + btn_h + gap;
-        let row2_w = bot_w * 4.0 + gap * 3.0;
-        let row2_x = (screen_w - row2_w) / 2.0;
-
-        if push_button(
-            &mut elements,
-            &mut any_hovered,
-            cursor,
-            row2_x,
-            row2_y,
-            bot_w,
-            btn_h,
-            gs,
-            "Edit",
-            has_sel,
-        ) && clicked
-        {
-            if let Some(idx) = self.selected_server {
-                if let Some(server) = self.server_list.servers.get(idx) {
-                    self.edit_name = server.name.clone();
-                    self.edit_address = server.address.clone();
-                    self.focused_field = Some(0);
-                    self.cursor_blink = Instant::now();
-                    self.screen = Screen::EditServer(idx);
-                }
-            }
-        }
-        if push_button(
-            &mut elements,
-            &mut any_hovered,
-            cursor,
-            row2_x + bot_w + gap,
-            row2_y,
-            bot_w,
-            btn_h,
-            gs,
-            "Delete",
-            has_sel,
-        ) && clicked
-        {
-            if let Some(idx) = self.selected_server {
-                self.screen = Screen::ConfirmDelete(idx);
-            }
-        }
-        if push_button(
-            &mut elements,
-            &mut any_hovered,
-            cursor,
-            row2_x + (bot_w + gap) * 2.0,
-            row2_y,
-            bot_w,
-            btn_h,
-            gs,
-            "Refresh",
-            true,
-        ) && clicked
-        {
-            self.refresh_servers();
-        }
-        if push_button(
-            &mut elements,
-            &mut any_hovered,
-            cursor,
-            row2_x + (bot_w + gap) * 3.0,
-            row2_y,
-            bot_w,
-            btn_h,
-            gs,
-            "Back",
-            true,
-        ) && clicked
-        {
-            self.screen = Screen::Main;
-        }
-
-        push_bottom_text(&mut elements, screen_w, screen_h, gs, text_width_fn);
-        MainMenuResult {
-            elements,
-            action,
-            cursor_pointer: any_hovered,
-            blur: 2.0,
-            clicked_button: false,
-        }
-    }
-
-    fn build_confirm_delete(
-        &mut self,
-        screen_w: f32,
-        screen_h: f32,
-        input: &MenuInput,
-        text_width_fn: &dyn Fn(&str, f32) -> f32,
-    ) -> MainMenuResult {
-        let Screen::ConfirmDelete(idx) = self.screen else {
-            return empty_result(2.0);
-        };
-
-        let gs = crate::ui::hud::gui_scale(screen_w, screen_h, self.gui_scale_setting);
-        let fs = common::FONT_SIZE * gs;
-        let form_w = FORM_W * gs;
-        let btn_h = common::BTN_H * gs;
-        let gap = BTN_GAP * gs;
-        let cursor = input.cursor;
-        let clicked = input.clicked;
-
-        if input.escape {
-            self.screen = Screen::ServerList;
-            return empty_result(2.0);
-        }
-
-        let warning = self
-            .server_list
-            .servers
-            .get(idx)
-            .map(|s| format!("'{}' will be lost forever! (A long time!)", s.name))
-            .unwrap_or_default();
-
-        let mut elements = Vec::new();
-        let mut any_hovered = false;
-
-        let cy = screen_h * 0.3;
-        elements.push(MenuElement::Text {
-            x: screen_w / 2.0,
-            y: cy,
-            text: "Are you sure?".into(),
-            scale: fs,
-            color: WHITE,
-            centered: true,
-        });
-        elements.push(MenuElement::Text {
-            x: screen_w / 2.0,
-            y: cy + fs + 12.0 * gs,
-            text: warning,
-            scale: fs,
-            color: COL_DIM,
-            centered: true,
-        });
-
-        let btn_x = (screen_w - form_w) / 2.0;
-        let btn_y = cy + fs * 2.0 + 44.0 * gs;
-
-        if push_button(
-            &mut elements,
-            &mut any_hovered,
-            cursor,
-            btn_x,
-            btn_y,
-            form_w,
-            btn_h,
-            gs,
-            "Delete",
-            true,
-        ) && clicked
-        {
-            self.server_list.remove(idx);
-            self.selected_server = None;
-            self.screen = Screen::ServerList;
-        }
-        if push_button(
-            &mut elements,
-            &mut any_hovered,
-            cursor,
-            btn_x,
-            btn_y + btn_h + gap,
-            form_w,
-            btn_h,
-            gs,
-            "Cancel",
-            true,
-        ) && clicked
-        {
-            self.screen = Screen::ServerList;
-        }
-
-        push_bottom_text(&mut elements, screen_w, screen_h, gs, text_width_fn);
-        MainMenuResult {
-            elements,
-            action: MenuAction::None,
-            cursor_pointer: any_hovered,
-            blur: 2.0,
-            clicked_button: false,
-        }
-    }
-
-    fn build_direct_connect(
-        &mut self,
-        screen_w: f32,
-        screen_h: f32,
-        input: &MenuInput,
-        text_width_fn: &dyn Fn(&str, f32) -> f32,
-    ) -> MainMenuResult {
-        let gs = crate::ui::hud::gui_scale(screen_w, screen_h, self.gui_scale_setting);
-        let fs = common::FONT_SIZE * gs;
-        let form_w = FORM_W * gs;
-        let btn_h = common::BTN_H * gs;
-        let gap = BTN_GAP * gs;
-        let field_h = FIELD_H * gs;
-        let cursor = input.cursor;
-        let clicked = input.clicked;
-
-        if input.escape {
-            self.screen = Screen::ServerList;
-            return empty_result(2.0);
-        }
-
-        self.handle_text_input(input, 1);
-
-        let mut elements = Vec::new();
-        let mut action = MenuAction::None;
-        let mut any_hovered = false;
-
-        let cx = screen_w / 2.0;
-        let form_x = cx - form_w / 2.0;
-        let mut y = 20.0 * gs;
-
-        elements.push(MenuElement::Text {
-            x: cx,
-            y,
-            text: "Direct Connect".into(),
-            scale: fs,
-            color: WHITE,
-            centered: true,
-        });
-        y += fs + 40.0 * gs;
-
-        elements.push(MenuElement::Text {
-            x: form_x,
-            y,
-            text: "Server Address".into(),
-            scale: fs,
-            color: COL_DIM,
-            centered: false,
-        });
-        y += fs + 4.0 * gs;
-
-        push_text_field(
-            &mut elements,
-            form_x,
-            y,
-            form_w,
-            field_h,
-            fs,
-            gs,
-            &self.edit_address,
-            self.focused_field == Some(0),
-            &self.cursor_blink,
-            text_width_fn,
-        );
-        if clicked && common::hit_test(cursor, [form_x, y, form_w, field_h]) {
-            self.focused_field = Some(0);
-            self.cursor_blink = Instant::now();
-        }
-        y += field_h + 28.0 * gs;
-
-        let valid = is_valid_address(&self.edit_address);
-        let enter_submit = input.enter && valid;
-
-        if (push_button(
-            &mut elements,
-            &mut any_hovered,
-            cursor,
-            form_x,
-            y,
-            form_w,
-            btn_h,
-            gs,
-            "Join Server",
-            valid,
-        ) && clicked)
-            || enter_submit
-        {
-            self.last_mp_ip = self.edit_address.clone();
-            action = MenuAction::Connect {
-                server: self.edit_address.clone(),
-                username: self.username.clone(),
-            };
-        }
-        y += btn_h + gap;
-        if push_button(
-            &mut elements,
-            &mut any_hovered,
-            cursor,
-            form_x,
-            y,
-            form_w,
-            btn_h,
-            gs,
-            "Cancel",
-            true,
-        ) && clicked
-        {
-            self.screen = Screen::ServerList;
-        }
-
-        push_bottom_text(&mut elements, screen_w, screen_h, gs, text_width_fn);
-        MainMenuResult {
-            elements,
-            action,
-            cursor_pointer: any_hovered,
-            blur: 2.0,
-            clicked_button: false,
-        }
-    }
-
-    fn build_edit_server(
-        &mut self,
-        screen_w: f32,
-        screen_h: f32,
-        input: &MenuInput,
-        text_width_fn: &dyn Fn(&str, f32) -> f32,
-    ) -> MainMenuResult {
-        let gs = crate::ui::hud::gui_scale(screen_w, screen_h, self.gui_scale_setting);
-        let fs = common::FONT_SIZE * gs;
-        let form_w = FORM_W * gs;
-        let btn_h = common::BTN_H * gs;
-        let gap = BTN_GAP * gs;
-        let field_h = FIELD_H * gs;
-        let cursor = input.cursor;
-        let clicked = input.clicked;
-
-        if input.escape {
-            self.screen = Screen::ServerList;
-            return empty_result(2.0);
-        }
-
-        self.handle_text_input(input, 2);
-
-        let mut elements = Vec::new();
-        let mut any_hovered = false;
-
-        let cx = screen_w / 2.0;
-        let form_x = cx - form_w / 2.0;
-        let mut y = 17.0 * gs;
-
-        elements.push(MenuElement::Text {
-            x: cx,
-            y,
-            text: "Edit Server Info".into(),
-            scale: fs,
-            color: WHITE,
-            centered: true,
-        });
-        y += fs + 20.0 * gs;
-
-        elements.push(MenuElement::Text {
-            x: form_x,
-            y,
-            text: "Server Name".into(),
-            scale: fs,
-            color: COL_DIM,
-            centered: false,
-        });
-        y += fs + 4.0 * gs;
-
-        push_text_field(
-            &mut elements,
-            form_x,
-            y,
-            form_w,
-            field_h,
-            fs,
-            gs,
-            &self.edit_name,
-            self.focused_field == Some(0),
-            &self.cursor_blink,
-            text_width_fn,
-        );
-        if clicked && common::hit_test(cursor, [form_x, y, form_w, field_h]) {
-            self.focused_field = Some(0);
-            self.cursor_blink = Instant::now();
-        }
-        y += field_h + 12.0 * gs;
-
-        elements.push(MenuElement::Text {
-            x: form_x,
-            y,
-            text: "Server Address".into(),
-            scale: fs,
-            color: COL_DIM,
-            centered: false,
-        });
-        y += fs + 4.0 * gs;
-
-        push_text_field(
-            &mut elements,
-            form_x,
-            y,
-            form_w,
-            field_h,
-            fs,
-            gs,
-            &self.edit_address,
-            self.focused_field == Some(1),
-            &self.cursor_blink,
-            text_width_fn,
-        );
-        if clicked && common::hit_test(cursor, [form_x, y, form_w, field_h]) {
-            self.focused_field = Some(1);
-            self.cursor_blink = Instant::now();
-        }
-        y += field_h + 28.0 * gs;
-
-        let valid = is_valid_address(&self.edit_address);
-        if push_button(
-            &mut elements,
-            &mut any_hovered,
-            cursor,
-            form_x,
-            y,
-            form_w,
-            btn_h,
-            gs,
-            "Done",
-            valid,
-        ) && clicked
-        {
-            let name = if self.edit_name.is_empty() {
-                "Minecraft Server".to_string()
-            } else {
-                self.edit_name.clone()
-            };
-            let addr = self.edit_address.clone();
-            let entry = ServerEntry {
-                name,
-                address: addr.clone(),
-            };
-            if let Screen::EditServer(idx) = self.screen {
-                self.server_list.update(idx, entry);
-            } else {
-                self.server_list.add(entry);
-            }
-            ping_all_servers(
-                &self.rt,
-                &[ServerEntry {
-                    name: String::new(),
-                    address: addr,
-                }],
-                &self.ping_results,
-            );
-            self.screen = Screen::ServerList;
-        }
-        y += btn_h + gap;
-        if push_button(
-            &mut elements,
-            &mut any_hovered,
-            cursor,
-            form_x,
-            y,
-            form_w,
-            btn_h,
-            gs,
-            "Cancel",
-            true,
-        ) && clicked
-        {
-            self.screen = Screen::ServerList;
-        }
-
-        push_bottom_text(&mut elements, screen_w, screen_h, gs, text_width_fn);
-        MainMenuResult {
-            elements,
-            action: MenuAction::None,
-            cursor_pointer: any_hovered,
-            blur: 2.0,
-            clicked_button: false,
-        }
-    }
-
-    fn handle_text_input(&mut self, input: &MenuInput, field_count: u8) {
-        if input.tab {
-            self.focused_field = Some(match self.focused_field {
-                Some(f) => (f + 1) % field_count,
-                None => 0,
-            });
-            self.cursor_blink = Instant::now();
-        }
-
-        let Some(field_idx) = self.focused_field else {
-            return;
-        };
-        let is_edit_form = matches!(self.screen, Screen::AddServer | Screen::EditServer(_));
-        let text = match (is_edit_form, field_idx) {
-            (true, 0) => &mut self.edit_name,
-            (true, 1) => &mut self.edit_address,
-            (false, 0) => &mut self.edit_address,
-            _ => return,
-        };
-
-        for ch in &input.typed_chars {
-            text.push(*ch);
-        }
-        if input.backspace {
-            text.pop();
-        }
-        if !input.typed_chars.is_empty() || input.backspace {
-            self.cursor_blink = Instant::now();
-        }
-    }
-
-    fn build_disconnected(
-        &mut self,
-        screen_w: f32,
-        screen_h: f32,
-        input: &MenuInput,
-        _text_width_fn: &dyn Fn(&str, f32) -> f32,
-    ) -> MainMenuResult {
-        let reason = match &self.screen {
-            Screen::Disconnected(r) => r.clone(),
-            _ => return empty_result(2.0),
-        };
-
-        let gs = crate::ui::hud::gui_scale(screen_w, screen_h, self.gui_scale_setting);
-        let title_size = 18.0 * gs;
-        let body_size = 11.0 * gs;
-        let btn_w = 160.0 * gs;
-        let btn_h = 30.0 * gs;
-        let gap = 12.0 * gs;
-
-        let cx = screen_w / 2.0;
-        let total_h = title_size + gap + body_size + gap * 2.0 + btn_h;
-        let top_y = (screen_h - total_h) / 2.0;
-
-        let mut elements = Vec::new();
-        let mut any_hovered = false;
-
-        elements.push(MenuElement::Text {
-            x: cx,
-            y: top_y,
-            text: "Disconnected".into(),
-            scale: title_size,
-            color: [1.0, 0.4, 0.4, 1.0],
-            centered: true,
-        });
-
-        elements.push(MenuElement::Text {
-            x: cx,
-            y: top_y + title_size + gap,
-            text: reason,
-            scale: body_size,
-            color: [0.85, 0.85, 0.85, 0.9],
-            centered: true,
-        });
-
-        let btn_y = top_y + title_size + gap + body_size + gap * 2.0;
-        if push_button(
-            &mut elements,
-            &mut any_hovered,
-            input.cursor,
-            cx - btn_w / 2.0,
-            btn_y,
-            btn_w,
-            btn_h,
-            gs,
-            "Back to Menu",
-            true,
-        ) && input.clicked
-        {
-            self.screen = Screen::Main;
-        }
-
-        MainMenuResult {
-            elements,
-            action: MenuAction::None,
-            cursor_pointer: any_hovered,
-            blur: 2.0,
-            clicked_button: false,
-        }
-    }
-
-    fn refresh_servers(&self) {
-        ping_all_servers(&self.rt, &self.server_list.servers, &self.ping_results);
-    }
-
-    fn build_options(&mut self, sw: f32, sh: f32, input: &MenuInput) -> MainMenuResult {
-        let fov_label = format!("FOV: {}", 70);
-        let rows: Vec<[&str; 2]> = vec![
-            [&fov_label, "Online"],
-            ["Skin Customization...", "Music & Sounds..."],
-            ["Video Settings...", "Controls..."],
-            ["Language...", "Chat Settings..."],
-            ["Resource Packs...", "Accessibility Settings..."],
-            ["Telemetry Data...", "Credits & Attribution..."],
-        ];
-
-        let nav: &[(&str, Screen)] = &[
-            ("Skin Customization...", Screen::OptionsSkinCustomization),
-            ("Music & Sounds...", Screen::OptionsMusicSounds),
-            ("Video Settings...", Screen::OptionsVideo),
-            ("Controls...", Screen::OptionsControls),
-            ("Language...", Screen::OptionsLanguage),
-            ("Chat Settings...", Screen::OptionsChatSettings),
-            ("Resource Packs...", Screen::OptionsResourcePacks),
-            ("Accessibility Settings...", Screen::OptionsAccessibility),
-            ("Telemetry Data...", Screen::OptionsTelemetry),
-            ("Credits & Attribution...", Screen::OptionsCredits),
-        ];
-
-        self.build_options_grid(sw, sh, input, "Options", Screen::Main, &rows, nav, &[])
-    }
-
-    fn build_options_video(&mut self, sw: f32, sh: f32, input: &MenuInput) -> MainMenuResult {
-        let fullscreen_label = match self.display_mode {
-            DisplayMode::Windowed => "Fullscreen: Windowed",
-            DisplayMode::Borderless => "Fullscreen: Borderless",
-            DisplayMode::Fullscreen => "Fullscreen: Exclusive",
-        };
-        let rd = format!("Render Distance: {} chunks", self.render_distance);
-        let sd = format!("Simulation Distance: {} chunks", self.render_distance);
-        let mf = format!("Max Framerate: {} fps", 120);
-        let gui_label = if self.gui_scale_setting == 0 {
-            "GUI Scale: Auto".to_string()
-        } else {
-            format!("GUI Scale: {}", self.gui_scale_setting)
-        };
-        let rows: Vec<[&str; 2]> = vec![
-            [&rd, &sd],
-            ["Graphics: Fancy", "Smooth Lighting: ON"],
-            [&mf, "VSync: OFF"],
-            ["View Bobbing: ON", &gui_label],
-            ["Attack Indicator: Crosshair", "Brightness: 50%"],
-            ["Clouds: Fancy", fullscreen_label],
-            ["Particles: All", "Mipmap Levels: 4"],
-        ];
-        let rd_frac = (self.render_distance as f32 - 2.0) / 30.0;
-        let sd_frac = (self.simulation_distance as f32 - 5.0) / 27.0;
-        let sliders: &[(&str, f32)] = &[
-            ("Render Distance:", rd_frac),
-            ("Simulation Distance:", sd_frac),
-        ];
-        self.build_options_grid(
-            sw,
-            sh,
-            input,
-            "Video Settings",
-            Screen::Options,
-            &rows,
-            &[],
-            sliders,
-        )
-    }
-
-    fn build_options_controls(&mut self, sw: f32, sh: f32, input: &MenuInput) -> MainMenuResult {
-        let rows: Vec<[&str; 2]> = vec![
-            ["Sensitivity: 100%", "Invert Mouse: OFF"],
-            ["Auto-Jump: ON", "Operator Items Tab: OFF"],
-            ["Key Binds...", "Mouse Settings..."],
-            ["Sneak: Toggle", "Sprint: Hold"],
-        ];
-        let nav: &[(&str, Screen)] = &[("Key Binds...", Screen::OptionsKeybinds)];
-        self.build_options_grid(sw, sh, input, "Controls", Screen::Options, &rows, nav, &[])
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn build_options_grid(
-        &mut self,
-        sw: f32,
-        sh: f32,
-        input: &MenuInput,
-        title: &str,
-        back: Screen,
-        rows: &[[&str; 2]],
-        nav: &[(&str, Screen)],
-        sliders: &[(&'static str, f32)],
-    ) -> MainMenuResult {
-        if input.escape {
-            self.screen = back.clone_screen();
-            return empty_result(2.0);
-        }
-
-        let gs = crate::ui::hud::gui_scale(sw, sh, self.gui_scale_setting);
-        let fs = common::FONT_SIZE * gs;
-        let btn_h = common::BTN_H * gs;
-        let gap = BTN_GAP * gs;
-        let header_h = HEADER_H * gs;
-        let sep_h = SEP_H * gs;
-        let btn_w = 150.0 * gs;
-        let half_w = (btn_w * 2.0 + gap) / 2.0;
-        let cx = sw / 2.0;
-        let cursor = input.cursor;
-        let clicked = input.clicked;
-
-        let mut elements = Vec::new();
-        let mut any_hovered = false;
-
-        common::push_overlay(&mut elements, sw, sh, 0.5);
-
-        elements.push(MenuElement::Text {
-            x: cx,
-            y: (header_h - fs) / 2.0,
-            text: title.into(),
-            scale: fs,
-            color: WHITE,
-            centered: true,
-        });
-        push_separator(&mut elements, 0.0, header_h, sw, sep_h);
-
-        let done_pad = 8.0 * gs;
-        let done_y = sh - btn_h - done_pad;
-        let content_top = header_h + sep_h;
-        let content_bottom = done_y;
-        let grid_h = rows.len() as f32 * btn_h + (rows.len() as f32 - 1.0).max(0.0) * gap;
-        let top_y = content_top + (content_bottom - content_top - grid_h) / 2.0;
-        let lx = cx - half_w;
-        let rx = lx + btn_w + gap;
-
-        let mut slider_results: Vec<(&str, f32)> = Vec::new();
-
-        for (row, pair) in rows.iter().enumerate() {
-            let by = top_y + row as f32 * (btn_h + gap);
-            for (col, label) in pair.iter().enumerate() {
-                let bx = if col == 0 { lx } else { rx };
-
-                if let Some((prefix, value)) = sliders.iter().find(|(p, _)| label.starts_with(p)) {
-                    let is_active = self.active_slider == Some(*prefix);
-                    let result = common::push_slider(
-                        &mut elements,
-                        cursor,
-                        input.mouse_held,
-                        bx,
-                        by,
-                        btn_w,
-                        btn_h,
-                        gs,
-                        fs,
-                        label,
-                        *value,
-                        is_active,
-                    );
-                    any_hovered |= result.hovered;
-                    if result.dragging {
-                        self.active_slider = Some(*prefix);
-                    }
-                    if let Some(v) = result.new_value {
-                        slider_results.push((prefix, v));
-                    }
-                    if !input.mouse_held && is_active {
-                        self.active_slider = None;
-                    }
-                    continue;
-                }
-
-                let h = common::push_button(
-                    &mut elements,
-                    cursor,
-                    bx,
-                    by,
-                    btn_w,
-                    btn_h,
-                    gs,
-                    fs,
-                    label,
-                    true,
-                );
-                any_hovered |= h;
-                if clicked && h {
-                    if let Some((_, target)) = nav.iter().find(|(l, _)| *l == *label) {
-                        self.screen = target.clone_screen();
-                    }
-                    if label.starts_with("GUI Scale:") {
-                        let max = crate::ui::hud::max_gui_scale(sw, sh);
-                        self.gui_scale_setting = (self.gui_scale_setting + 1) % (max + 1);
-                        self.save_settings();
-                    }
-                    if label.starts_with("Fullscreen:") {
-                        self.display_mode = self.display_mode.cycle();
-                    }
-                }
-            }
-        }
-
-        for (prefix, value) in &slider_results {
-            if *prefix == "Render Distance:" {
-                self.render_distance = (2.0 + value * 30.0).round() as u32;
-                self.save_settings();
-            }
-            if *prefix == "Simulation Distance:" {
-                self.simulation_distance = (5.0 + value * 27.0).round() as u32;
-                self.save_settings();
-            }
-        }
-
-        let done_w = btn_w * 2.0 + gap;
-        let h = common::push_button(
-            &mut elements,
-            cursor,
-            cx - done_w / 2.0,
-            done_y,
-            done_w,
-            btn_h,
-            gs,
-            fs,
-            "Done",
-            true,
-        );
-        any_hovered |= h;
-        if clicked && h {
-            self.screen = back;
-        }
-
-        MainMenuResult {
-            elements,
-            action: MenuAction::None,
-            cursor_pointer: any_hovered,
-            blur: 2.0,
-            clicked_button: false,
-        }
-    }
-
-    fn build_options_stub(
-        &mut self,
-        sw: f32,
-        sh: f32,
-        input: &MenuInput,
-        title: &str,
-        back: Screen,
-    ) -> MainMenuResult {
-        if input.escape {
-            self.screen = back.clone_screen();
-            return empty_result(2.0);
-        }
-
-        let gs = crate::ui::hud::gui_scale(sw, sh, self.gui_scale_setting);
-        let fs = common::FONT_SIZE * gs;
-        let btn_h = common::BTN_H * gs;
-        let gap = BTN_GAP * gs;
-        let header_h = HEADER_H * gs;
-        let sep_h = SEP_H * gs;
-        let cx = sw / 2.0;
-
-        let mut elements = Vec::new();
-        let mut any_hovered = false;
-
-        common::push_overlay(&mut elements, sw, sh, 0.5);
-
-        elements.push(MenuElement::Text {
-            x: cx,
-            y: (header_h - fs) / 2.0,
-            text: title.into(),
-            scale: fs,
-            color: WHITE,
-            centered: true,
-        });
-        push_separator(&mut elements, 0.0, header_h, sw, sep_h);
-
-        let body_fs = 10.0 * gs;
-        elements.push(MenuElement::Text {
-            x: cx,
-            y: sh / 2.0 - body_fs,
-            text: "Coming soon".into(),
-            scale: body_fs,
-            color: COL_DIM,
-            centered: true,
-        });
-
-        let done_w = 150.0 * gs * 2.0 + gap;
-        let done_y = sh - btn_h - 8.0 * gs;
-        let h = common::push_button(
-            &mut elements,
-            input.cursor,
-            cx - done_w / 2.0,
-            done_y,
-            done_w,
-            btn_h,
-            gs,
-            fs,
-            "Done",
-            true,
-        );
-        any_hovered |= h;
-        if input.clicked && h {
-            self.screen = back;
-        }
-
-        MainMenuResult {
-            elements,
-            action: MenuAction::None,
-            cursor_pointer: any_hovered,
-            blur: 2.0,
-            clicked_button: false,
-        }
-    }
-}
-
-fn empty_result(blur: f32) -> MainMenuResult {
-    MainMenuResult {
-        elements: Vec::new(),
-        action: MenuAction::None,
-        cursor_pointer: false,
-        blur,
-        clicked_button: false,
-    }
-}
-
-fn push_separator(elements: &mut Vec<MenuElement>, x: f32, y: f32, w: f32, h: f32) {
-    elements.push(MenuElement::Rect {
-        x,
-        y,
-        w,
-        h,
-        corner_radius: 0.0,
-        color: COL_SEP,
-    });
-}
-
-fn push_outline(elements: &mut Vec<MenuElement>, x: f32, y: f32, w: f32, h: f32, gs: f32) {
-    let t = 1.0 * gs;
-    let c = WHITE;
-    elements.push(MenuElement::Rect {
-        x,
-        y,
-        w,
-        h: t,
-        corner_radius: 0.0,
-        color: c,
-    });
-    elements.push(MenuElement::Rect {
-        x,
-        y: y + h - t,
-        w,
-        h: t,
-        corner_radius: 0.0,
-        color: c,
-    });
-    elements.push(MenuElement::Rect {
-        x,
-        y: y + t,
-        w: t,
-        h: h - t * 2.0,
-        corner_radius: 0.0,
-        color: c,
-    });
-    elements.push(MenuElement::Rect {
-        x: x + w - t,
-        y: y + t,
-        w: t,
-        h: h - t * 2.0,
-        corner_radius: 0.0,
-        color: c,
-    });
-}
-
-#[allow(clippy::too_many_arguments)]
-fn push_button(
-    elements: &mut Vec<MenuElement>,
-    any_hovered: &mut bool,
-    cursor: (f32, f32),
-    x: f32,
-    y: f32,
-    w: f32,
-    h: f32,
-    gs: f32,
-    label: &str,
-    enabled: bool,
-) -> bool {
-    let hovered = common::push_button(
-        elements,
-        cursor,
-        x,
-        y,
-        w,
-        h,
-        gs,
-        common::FONT_SIZE * gs,
-        label,
-        enabled,
-    );
-    *any_hovered |= hovered;
-    hovered
-}
-
-#[allow(clippy::too_many_arguments)]
-fn push_text_field(
-    elements: &mut Vec<MenuElement>,
-    x: f32,
-    y: f32,
-    w: f32,
-    h: f32,
-    fs: f32,
-    gs: f32,
-    text: &str,
-    focused: bool,
-    cursor_blink: &Instant,
-    text_width_fn: &dyn Fn(&str, f32) -> f32,
-) {
-    let border = if focused {
-        FIELD_BORDER_FOCUS
-    } else {
-        FIELD_BORDER
-    };
-    elements.push(MenuElement::Rect {
-        x: x - gs,
-        y: y - gs,
-        w: w + gs * 2.0,
-        h: h + gs * 2.0,
-        corner_radius: 0.0,
-        color: border,
-    });
-    elements.push(MenuElement::Rect {
-        x,
-        y,
-        w,
-        h,
-        corner_radius: 0.0,
-        color: FIELD_BG,
-    });
-
-    let pad = 4.0 * gs;
-    elements.push(MenuElement::Text {
-        x: x + pad,
-        y: y + (h - fs) / 2.0,
-        text: text.into(),
-        scale: fs,
-        color: WHITE,
-        centered: false,
-    });
-
-    if focused {
-        let text_w = text_width_fn(text, fs);
-        common::push_cursor_blink(
-            elements,
-            cursor_blink,
-            x + pad,
-            y + (h - fs) / 2.0,
-            gs,
-            fs,
-            text_w,
-        );
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-fn push_server_status(
-    elements: &mut Vec<MenuElement>,
-    ping_results: &std::collections::HashMap<String, PingState>,
-    address: &str,
-    text_x: f32,
-    motd_y: f32,
-    entry_rect: &[f32; 4],
-    fs: f32,
-    gs: f32,
-    text_width_fn: &dyn Fn(&str, f32) -> f32,
-) {
-    let Some(state) = ping_results.get(address) else {
-        elements.push(MenuElement::Text {
-            x: text_x,
-            y: motd_y,
-            text: address.into(),
-            scale: fs,
-            color: COL_DARK_DIM,
-            centered: false,
-        });
-        return;
-    };
-
-    match state {
-        PingState::Pinging => {
-            let dots = match (std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .unwrap_or_default()
-                .as_millis()
-                / 500)
-                % 4
-            {
-                0 => "Pinging",
-                1 => "Pinging.",
-                2 => "Pinging..",
-                _ => "Pinging...",
-            };
-            elements.push(MenuElement::Text {
-                x: text_x,
-                y: motd_y,
-                text: dots.into(),
-                scale: fs,
-                color: COL_DARK_DIM,
-                centered: false,
-            });
-        }
-        PingState::Success {
-            motd,
-            online,
-            max,
-            latency_ms,
-            ..
-        } => {
-            elements.push(MenuElement::McText {
-                x: text_x,
-                y: motd_y,
-                spans: motd.clone(),
-                scale: fs,
-                centered: false,
-            });
-
-            let player_text = format!("{online}/{max}");
-            let right_x = entry_rect[0] + entry_rect[2] - 10.0 * gs;
-            let pw = text_width_fn(&player_text, fs);
-            elements.push(MenuElement::Text {
-                x: right_x - pw,
-                y: entry_rect[1] + 1.0 * gs,
-                text: player_text,
-                scale: fs,
-                color: COL_DARK_DIM,
-                centered: false,
-            });
-
-            let (bars, bar_color) = ping_level(*latency_ms);
-            let bw = 10.0 * gs;
-            let bh = 8.0 * gs;
-            let bx = right_x - pw - 6.0 * gs - bw;
-            let by = entry_rect[1] + 1.0 * gs;
-            push_ping_bars(elements, bx, by, bw, bh, bars, bar_color);
-        }
-        PingState::Failed(err) => {
-            let display = if err.len() > 40 {
-                "Can't connect to server"
-            } else {
-                err
-            };
-            elements.push(MenuElement::Text {
-                x: text_x,
-                y: motd_y,
-                text: display.into(),
-                scale: fs,
-                color: COL_RED,
-                centered: false,
-            });
-        }
-    }
-}
-
-const PING_THRESHOLDS: [(u64, u8, [f32; 4]); 5] = [
-    (150, 5, [0.26, 0.63, 0.28, 1.0]),
-    (300, 4, [0.51, 0.78, 0.52, 1.0]),
-    (600, 3, [1.0, 0.93, 0.35, 1.0]),
-    (1000, 2, [1.0, 0.65, 0.15, 1.0]),
-    (u64::MAX, 1, [0.9, 0.22, 0.21, 1.0]),
-];
-
-fn ping_level(ms: u64) -> (u8, [f32; 4]) {
-    for &(threshold, bars, color) in &PING_THRESHOLDS {
-        if ms < threshold {
-            return (bars, color);
-        }
-    }
-    (1, PING_THRESHOLDS[4].2)
-}
-
-fn push_ping_bars(
-    elements: &mut Vec<MenuElement>,
-    x: f32,
-    y: f32,
-    w: f32,
-    h: f32,
-    bars: u8,
-    color: [f32; 4],
-) {
-    let bw = w / 5.0;
-    let inactive = [0.12, 0.12, 0.16, 1.0];
-    for i in 0..5u8 {
-        let bh = h * (i as f32 + 1.0) / 5.0;
-        let bx = x + i as f32 * bw;
-        let by = y + h - bh;
-        elements.push(MenuElement::Rect {
-            x: bx,
-            y: by,
-            w: bw - 1.0,
-            h: bh,
-            corner_radius: 0.0,
-            color: if i < bars { color } else { inactive },
-        });
-    }
-}
-
-fn push_bottom_text(
-    elements: &mut Vec<MenuElement>,
-    screen_w: f32,
-    screen_h: f32,
-    gs: f32,
-    text_width_fn: &dyn Fn(&str, f32) -> f32,
-) {
-    let fs = 7.0 * gs;
-    let pad = 4.0 * gs;
-    let y = screen_h - pad - fs;
-    let col = [0.39, 0.55, 0.78, 0.3];
-
-    elements.push(MenuElement::Text {
-        x: pad,
-        y,
-        text: "Minecraft 1.21.11".into(),
-        scale: fs,
-        color: col,
-        centered: false,
-    });
-
-    let name = "POMC";
-    let tag = "early dev";
-    let tag_size = fs * 0.65;
-    let gap = 2.0 * gs;
-    let nw = text_width_fn(name, fs);
-    let tw = text_width_fn(tag, tag_size);
-    let nx = screen_w - pad - nw - gap - tw;
-    elements.push(MenuElement::Text {
-        x: nx,
-        y,
-        text: name.into(),
-        scale: fs,
-        color: col,
-        centered: false,
-    });
-    elements.push(MenuElement::Text {
-        x: nx + nw + gap,
-        y,
-        text: tag.into(),
-        scale: tag_size,
-        color: col,
-        centered: false,
-    });
-}
-
-struct DropdownStyle {
-    item_h: f32,
-    radius: f32,
-    font: f32,
-    icon_scale: f32,
-    pad: f32,
-}
-
-impl DropdownStyle {
-    fn new(gs: f32) -> Self {
-        Self {
-            item_h: 28.0 * gs,
-            radius: 5.0 * gs,
-            font: 9.0 * gs,
-            icon_scale: 11.0 * gs,
-            pad: 10.0 * gs,
-        }
-    }
-
-    fn draw_background(&self, elements: &mut Vec<MenuElement>, x: f32, y: f32, w: f32, h: f32) {
-        elements.push(MenuElement::Rect {
-            x,
-            y,
-            w,
-            h,
-            corner_radius: self.radius,
-            color: [0.08, 0.08, 0.12, 0.92],
-        });
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn draw_item(
-        &self,
-        elements: &mut Vec<MenuElement>,
-        any_hovered: &mut bool,
-        cursor: (f32, f32),
-        drop_x: f32,
-        drop_y: f32,
-        drop_w: f32,
-        index: usize,
-        count: usize,
-        label: &str,
-        icon: Option<(char, [f32; 4])>,
-        hover_color: [f32; 4],
-        normal_color: [f32; 4],
-    ) -> bool {
-        let iy = drop_y + index as f32 * self.item_h;
-        let rect = [drop_x, iy, drop_w, self.item_h];
-        let hovered = common::hit_test(cursor, rect);
-        *any_hovered |= hovered;
-
-        if hovered {
-            let r = if index == 0 || index == count - 1 {
-                self.radius
-            } else {
-                0.0
-            };
-            elements.push(MenuElement::Rect {
-                x: drop_x,
-                y: iy,
-                w: drop_w,
-                h: self.item_h,
-                corner_radius: r,
-                color: [1.0, 1.0, 1.0, 0.08],
-            });
-        }
-
-        if let Some((icon_char, icon_col)) = icon {
-            elements.push(MenuElement::Icon {
-                x: drop_x + self.pad + self.icon_scale / 2.0,
-                y: iy + self.item_h / 2.0,
-                icon: icon_char,
-                scale: self.icon_scale,
-                color: if hovered { hover_color } else { icon_col },
-            });
-        }
-
-        elements.push(MenuElement::Text {
-            x: drop_x + self.pad + self.icon_scale + 6.0,
-            y: iy + (self.item_h - self.font) / 2.0,
-            text: label.to_string(),
-            scale: self.font,
-            color: if hovered { hover_color } else { normal_color },
-            centered: false,
-        });
-
-        hovered
-    }
-}
-
-fn ease_out_cubic(t: f32) -> f32 {
-    let t = 1.0 - t;
-    1.0 - t * t * t
-}
-
-fn dismiss_dropdown(
-    cursor: (f32, f32),
-    clicked: bool,
-    clicked_inside: bool,
-    dropdown: [f32; 4],
-    anchor: [f32; 4],
-) -> bool {
-    clicked
-        && !clicked_inside
-        && !common::hit_test(cursor, dropdown)
-        && !common::hit_test(cursor, anchor)
-}
-
-fn smoothstep(t: f32) -> f32 {
-    t * t * (3.0 - 2.0 * t)
-}
-
-fn emit_transition_strips(
-    elements: &mut Vec<MenuElement>,
-    screen_w: f32,
-    screen_h: f32,
-    close_t: f32,
-    open_t: f32,
-) {
-    let strip_w = screen_w / STRIP_COUNT as f32 + 1.0;
-    let strip_h = screen_h * 2.0;
-    let wave_spread = 0.3;
-    for i in 0..STRIP_COUNT {
-        let fi = i as f32 / STRIP_COUNT as f32;
-        let close_ease =
-            smoothstep(((close_t - fi * wave_spread) / (1.0 - wave_spread)).clamp(0.0, 1.0));
-        let ri = (STRIP_COUNT - 1 - i) as f32 / STRIP_COUNT as f32;
-        let open_ease =
-            smoothstep(((open_t - ri * wave_spread) / (1.0 - wave_spread)).clamp(0.0, 1.0));
-        let y = -strip_h + close_ease * screen_h - open_ease * screen_h;
-        let sx = i as f32 * (strip_w - 1.0);
-        let hue_shift = fi * 0.08;
-        elements.push(MenuElement::Rect {
-            x: sx,
-            y,
-            w: strip_w,
-            h: strip_h,
-            corner_radius: 0.0,
-            color: [0.04 + hue_shift, 0.02, 0.12 + hue_shift * 0.5, 1.0],
-        });
-        elements.push(MenuElement::Rect {
-            x: sx,
-            y,
-            w: 1.0,
-            h: strip_h,
-            corner_radius: 0.0,
-            color: [0.3, 0.15, 0.5, 0.3 * (1.0 - open_ease)],
-        });
-    }
-}
